refactor: streamline services and fix async overhead
