+++ backend/PhotoBank.Services/Enrichers/AdultEnricher.cs
@@ -11,15 +11,13 @@ namespace PhotoBank.Services.Enrichers
 
         public Type[] Dependencies => new Type[1] { typeof(AnalyzeEnricher) };
 
-        public async Task EnrichAsync(Photo photo, SourceDataDto sourceData)
+        public Task EnrichAsync(Photo photo, SourceDataDto sourceData)
         {
-            await Task.Run(() =>
-            {
-                photo.IsAdultContent = sourceData.ImageAnalysis.Adult.IsAdultContent;
-                photo.AdultScore = sourceData.ImageAnalysis.Adult.AdultScore;
-                photo.IsRacyContent = sourceData.ImageAnalysis.Adult.IsRacyContent;
-                photo.RacyScore = sourceData.ImageAnalysis.Adult.RacyScore;
-            });
+            photo.IsAdultContent = sourceData.ImageAnalysis.Adult.IsAdultContent;
+            photo.AdultScore = sourceData.ImageAnalysis.Adult.AdultScore;
+            photo.IsRacyContent = sourceData.ImageAnalysis.Adult.IsRacyContent;
+            photo.RacyScore = sourceData.ImageAnalysis.Adult.RacyScore;
+            return Task.CompletedTask;
         }
     }
 }
