+++ backend/PhotoBank.Services/Enrichers/CaptionEnricher.cs
@@ -11,20 +11,19 @@ namespace PhotoBank.Services.Enrichers
         public EnricherType EnricherType => EnricherType.Caption;
         public Type[] Dependencies => new Type[1] { typeof(AnalyzeEnricher) };
 
-        public async Task EnrichAsync(Photo photo, SourceDataDto sourceData)
+        public Task EnrichAsync(Photo photo, SourceDataDto sourceData)
         {
-            await Task.Run(() =>
+            photo.Captions = new List<Caption>();
+            foreach (var caption in sourceData.ImageAnalysis.Description.Captions)
             {
-                photo.Captions = new List<Caption>();
-                foreach (var caption in sourceData.ImageAnalysis.Description.Captions)
+                photo.Captions.Add(new Caption
                 {
-                    photo.Captions.Add(new Caption
-                    {
-                        Confidence = caption.Confidence,
-                        Text = caption.Text
-                    });
-                }
-            });
+                    Confidence = caption.Confidence,
+                    Text = caption.Text
+                });
+            }
+
+            return Task.CompletedTask;
         }
     }
 }
