+++ backend/PhotoBank.Services/Enrichers/MetadataEnricher.cs
@@ -26,50 +26,53 @@ namespace PhotoBank.Services.Enrichers
         public EnricherType EnricherType => EnricherType.Metadata;
         public Type[] Dependencies => new Type[1] { typeof(PreviewEnricher) };
 
-        public async Task EnrichAsync(Photo photo, SourceDataDto sourceData)
+        public Task EnrichAsync(Photo photo, SourceDataDto sourceData)
         {
-            await Task.Run(() =>
+            var normalizedAbsolutePath = sourceData.AbsolutePath.Replace('\\', Path.DirectorySeparatorChar).Replace('/', Path.DirectorySeparatorChar);
+            var normalizedStoragePath = photo.Storage.Folder.Replace('\\', Path.DirectorySeparatorChar).Replace('/', Path.DirectorySeparatorChar);
+
+            photo.Name = Path.GetFileNameWithoutExtension(normalizedAbsolutePath);
+            photo.RelativePath = Path.GetDirectoryName(normalizedAbsolutePath)?
+                .Replace(normalizedStoragePath, string.Empty)
+                .TrimStart(Path.DirectorySeparatorChar);
+            photo.Files = new List<File>
             {
-                photo.Name = Path.GetFileNameWithoutExtension(sourceData.AbsolutePath);
-                photo.RelativePath =
-                    Path.GetDirectoryName(Path.GetRelativePath(photo.Storage.Folder, sourceData.AbsolutePath));
-                photo.Files = new List<File>
+                new()
                 {
-                        new()
-                        {
-                            Name = Path.GetFileName(sourceData.AbsolutePath)
-                        }
-                };
-
-                IEnumerable<Directory> directories = _imageMetadataReaderWrapper.ReadMetadata(sourceData.AbsolutePath);
-
-                var exifIfd0Directory = directories.OfType<ExifIfd0Directory>().FirstOrDefault();
-                var exifSubIfdDirectory = directories.OfType<ExifSubIfdDirectory>().FirstOrDefault();
-                var fileMetadataDirectory = directories.OfType<FileMetadataDirectory>().FirstOrDefault();
-
-                var gpsDirectory = directories.OfType<GpsDirectory>().FirstOrDefault();
-
-                if (exifSubIfdDirectory != null || exifIfd0Directory != null || fileMetadataDirectory != null)
-                {
-                    photo.TakenDate = GetTakenDate(new Directory[] { exifIfd0Directory, exifSubIfdDirectory, fileMetadataDirectory });
+                    Name = Path.GetFileName(normalizedAbsolutePath)
                 }
+            };
 
-                if (exifSubIfdDirectory != null)
-                {
-                    photo.Height ??= GetHeight(exifSubIfdDirectory);
-                    photo.Width ??= GetWidth(exifSubIfdDirectory);
-                }
+            IEnumerable<Directory> directories = _imageMetadataReaderWrapper.ReadMetadata(sourceData.AbsolutePath);
 
-                if (exifIfd0Directory != null)
-                {
-                    photo.Orientation ??= GetOrientation(exifIfd0Directory);
-                }
+            var exifIfd0Directory = directories.OfType<ExifIfd0Directory>().FirstOrDefault();
+            var exifSubIfdDirectory = directories.OfType<ExifSubIfdDirectory>().FirstOrDefault();
+            var fileMetadataDirectory = directories.OfType<FileMetadataDirectory>().FirstOrDefault();
 
-                if (gpsDirectory != null)
-                {
-                    photo.Location = GeoWrapper.GetLocation(gpsDirectory);
-                }
-            });
+            var gpsDirectory = directories.OfType<GpsDirectory>().FirstOrDefault();
+
+            if (exifSubIfdDirectory != null || exifIfd0Directory != null || fileMetadataDirectory != null)
+            {
+                photo.TakenDate = GetTakenDate(new Directory[] { exifIfd0Directory, exifSubIfdDirectory, fileMetadataDirectory });
+            }
+
+            if (exifSubIfdDirectory != null)
+            {
+                photo.Height ??= GetHeight(exifSubIfdDirectory);
+                photo.Width ??= GetWidth(exifSubIfdDirectory);
+            }
+
+            if (exifIfd0Directory != null)
+            {
+                photo.Orientation ??= GetOrientation(exifIfd0Directory);
+            }
+
+            if (gpsDirectory != null)
+            {
+                photo.Location = GeoWrapper.GetLocation(gpsDirectory);
+            }
+
+            return Task.CompletedTask;
         }
 
         private static DateTime? GetTakenDate(IEnumerable<Directory> directories)
