+++ backend/PhotoBank.Services/Enrichers/ObjectPropertyEnricher.cs
